refactor: 모듈 간 경계 개선을 위해 manualHandlerMapping 생성을 config로 분리
- ManualMappingConfig는 app 모듈에서 app 모듈의 컨트롤러 import.
- ManualHandlerMapping을 mvc 모듈로 옮길 수 있음.
