+++ app/src/main/java/com/techcourse/ManualMappingConfig.java
@@ -1,0 +1,26 @@
+package com.techcourse;
+
+import com.interface21.webmvc.servlet.mapping.ManualHandlerMapping;
+import com.interface21.webmvc.servlet.mvc.asis.Controller;
+import com.interface21.webmvc.servlet.mvc.asis.ForwardController;
+import com.techcourse.controller.LoginController;
+import com.techcourse.controller.LoginViewController;
+import com.techcourse.controller.LogoutController;
+import java.util.HashMap;
+import java.util.Map;
+import lombok.AccessLevel;
+import lombok.NoArgsConstructor;
+
+@NoArgsConstructor(access = AccessLevel.NONE)
+public class ManualMappingConfig {
+
+    public static ManualHandlerMapping createManualMapping() {
+        final Map<String, Controller> controllers = new HashMap<>();
+        controllers.put("/", new ForwardController("/index.jsp"));
+        controllers.put("/login", new LoginController());
+        controllers.put("/login/view", new LoginViewController());
+        controllers.put("/logout", new LogoutController());
+
+        return ManualHandlerMapping.from(controllers);
+    }
+}
