+++ mvc/src/main/java/com/interface21/webmvc/servlet/mapping/ManualHandlerMapping.java
@@ -1,13 +1,7 @@
-package com.techcourse;
+package com.interface21.webmvc.servlet.mapping;
 
-import com.interface21.webmvc.servlet.mapping.HandlerMapping;
 import com.interface21.webmvc.servlet.mvc.asis.Controller;
-import com.interface21.webmvc.servlet.mvc.asis.ForwardController;
-import com.techcourse.controller.LoginController;
-import com.techcourse.controller.LoginViewController;
-import com.techcourse.controller.LogoutController;
 import jakarta.servlet.http.HttpServletRequest;
-import java.util.HashMap;
 import java.util.Map;
 import lombok.AccessLevel;
 import lombok.RequiredArgsConstructor;
@@ -19,13 +13,7 @@ public class ManualHandlerMapping implements HandlerMapping {
 
     private final Map<String, Controller> controllers;
 
-    public static ManualHandlerMapping initialize() {
-        final Map<String, Controller> controllers = new HashMap<>();
-        controllers.put("/", new ForwardController("/index.jsp"));
-        controllers.put("/login", new LoginController());
-        controllers.put("/login/view", new LoginViewController());
-        controllers.put("/logout", new LogoutController());
-
+    public static ManualHandlerMapping from(final Map<String, Controller> controllers) {
         log.info("Initialized Handler Mapping!");
         controllers.keySet()
                 .forEach(path -> log.info("Path : {}, Controller : {}", path, controllers.get(path).getClass()));
